requirements.txt
streamlit>=1.25.0  # chat_input や chat_message に対応
openai>=1.2.4
python-dotenv
pyyaml>=6.0
